// src/types.ts
export interface User {
  name: string;
  role: 'admin' | 'contributor';
}

export type IssueStatus = 'Backlog' | 'In Progress' | 'Done';
export type IssuePriority = 'low' | 'medium' | 'high';

export interface Issue {
  id: string;
  title: string;
  status: IssueStatus;
  priority: IssuePriority;
  severity: number;
  createdAt: string;
  assignee: string;
  tags: string[];
  userDefinedRank?: number;
}

export interface ErrorResponse {
  message: string;
  status: number;
}

export interface Filter {
  search: string;
  assignee: string;
  severity: string;
}